+++ workspace/ks-core/src/compiler_new/types.rs
@@ -1,4 +1,4 @@
-pub type VariableId = u64;
+pub type VariableId = u32;
 pub type Pointer = usize;
 pub type CollectionId = usize;
 pub type NativeId = usize;
